Moved locationMarker to be a separate file and mocked it in testing
